move align_types
