+++ src/frontend/src/binder/expr/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use itertools::{zip_eq, Itertools as _};
+use itertools::zip_eq;
 use risingwave_common::error::{ErrorCode, Result};
 use risingwave_common::types::DataType;
 use risingwave_sqlparser::ast::{
@@ -20,7 +20,7 @@ use risingwave_sqlparser::ast::{
 };
 
 use crate::binder::Binder;
-use crate::expr::{least_restrictive, Expr as _, ExprImpl, ExprType, FunctionCall, SubqueryKind};
+use crate::expr::{Expr as _, ExprImpl, ExprType, FunctionCall, SubqueryKind};
 
 mod binary_op;
 mod column;
@@ -313,30 +313,4 @@ pub fn bind_data_type(data_type: &AstDataType) -> Result<DataType> {
         }
     };
     Ok(data_type)
-}
-
-/// Find the `least_restrictive` type over a list of `exprs`, and add implicit cast when necessary.
-/// Used by `VALUES`, `CASE`, `UNION`, etc. See [PG](https://www.postgresql.org/docs/current/typeconv-union-case.html).
-pub fn align_types<'a>(exprs: impl Iterator<Item = &'a mut ExprImpl>) -> Result<DataType> {
-    use std::mem::swap;
-
-    let exprs = exprs.collect_vec();
-    // Essentially a filter_map followed by a try_reduce, which is unstable.
-    let mut ret_type = None;
-    for e in &exprs {
-        if e.is_null() {
-            continue;
-        }
-        ret_type = match ret_type {
-            None => Some(e.return_type()),
-            Some(t) => Some(least_restrictive(t, e.return_type())?),
-        };
-    }
-    let ret_type = ret_type.unwrap_or(DataType::Varchar);
-    for e in exprs {
-        let mut dummy = ExprImpl::literal_bool(false);
-        swap(&mut dummy, e);
-        *e = dummy.cast_implicit(ret_type.clone())?;
-    }
-    Ok(ret_type)
 }
